+++ empymod/model.py
@@ -1490,6 +1490,15 @@ def fem(ab, off, angle, zsrc, zrec, lsrc, lrec, depth, freq, etaH, etaV, zetaH,
     # Get angle dependent factors
     factAng = kernel.angle_factor(angle, ab, msrc, mrec)
 
+    # Compute required lambdas for given hankel-filter-base
+    # This should be in utils, but this is a backwards-incompatible change.
+    # Move this to utils for version 2.0.
+    if ht == 'fht':
+        # htarg[0] = filter; htarg[1] = pts_per_dec
+        lambd, int_pts = transform.get_spline_values(htarg[0], off, htarg[1])
+        if not loop_off:
+            htarg = (htarg[0], htarg[1], lambd, int_pts)
+
     # If not full-space with xdirect calculate fEM-field
     if not isfullspace*xdir:
         calc = getattr(transform, ht)
@@ -1506,6 +1515,12 @@ def fem(ab, off, angle, zsrc, zrec, lsrc, lrec, depth, freq, etaH, etaV, zetaH,
 
         elif loop_off:
             for i in range(off.size):
+
+                # See comments above where it says "ht == 'fht'".
+                # Get pre-calculated lambd, int_pts for this offset
+                if ht == 'fht':
+                    htarg = (htarg[0], htarg[1], lambd[None, i, :], int_pts[i])
+
                 out = calc(zsrc, zrec, lsrc, lrec, off[None, i],
                            factAng[None, i], depth, ab, etaH, etaV, zetaH,
                            zetaV, xdir, htarg, use_ne_eval, msrc, mrec)